requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.10.4
huggingface-hub==0.23.5
datasets==2.20.0
openai==1.57.2
httpx==0.27.2
nest_asyncio==1.5.8
PyMuPDF==1.25.1
python-docx==1.1.2
#better-profanity==0.7.0
#guardrails-ai==0.6.1

# AWS SDK
boto3==1.35.48
botocore==1.35.48
awscli

# API and CORS
# starlette>=0.40.0
# python-multipart==0.0.9

# Type hints and annotations
typing-extensions==4.12.2

# Environment variables
python-dotenv==1.0.0



# Optional: Logging and monitoring
loguru==0.7.2

#frontend
#streamlit==1.39.0

# Note: Install Git LFS manually from https://git-lfs.github.com/ and run `git lfs install`.

